+++ stac_fastapi/api/stac_fastapi/api/app.py
@@ -62,14 +62,21 @@ class StacApi:
     exceptions: Dict[Type[Exception], int] = attr.ib(
         default=attr.Factory(lambda: DEFAULT_STATUS_CODES)
     )
-    app: FastAPI = attr.ib(default=attr.Factory(FastAPI))
+    app: FastAPI = attr.ib(
+        default=attr.Factory(
+            lambda self: FastAPI(openapi_url=self.settings.openapi_url), takes_self=True
+        )
+    )
     router: APIRouter = attr.ib(default=attr.Factory(APIRouter))
     title: str = attr.ib(default="stac-fastapi")
     api_version: str = attr.ib(default="0.1")
     stac_version: str = attr.ib(default=STAC_VERSION)
     description: str = attr.ib(default="stac-fastapi")
     search_request_model: Type[Search] = attr.ib(default=STACSearch)
+    search_get_request: Type[SearchGetRequest] = attr.ib(default=SearchGetRequest)
+    item_collection_uri: Type[ItemCollectionUri] = attr.ib(default=ItemCollectionUri)
     response_class: Type[Response] = attr.ib(default=JSONResponse)
+    middlewares: List = attr.ib(default=attr.Factory(lambda: [BrotliMiddleware]))
 
     def get_extension(self, extension: Type[ApiExtension]) -> Optional[ApiExtension]:
         """Get an extension.
@@ -194,7 +201,9 @@ class StacApi:
             response_model_exclude_unset=True,
             response_model_exclude_none=True,
             methods=["GET"],
-            endpoint=self._create_endpoint(self.client.get_search, SearchGetRequest),
+            endpoint=self._create_endpoint(
+                self.client.get_search, self.search_get_request
+            ),
         )
 
     def register_get_collections(self):
@@ -250,7 +259,7 @@ class StacApi:
             response_model_exclude_none=True,
             methods=["GET"],
             endpoint=self._create_endpoint(
-                self.client.item_collection, ItemCollectionUri
+                self.client.item_collection, self.item_collection_uri
             ),
         )
 
@@ -341,5 +350,6 @@ class StacApi:
         # customize openapi
         self.app.openapi = self.customize_openapi
 
-        # add compression middleware
-        self.app.add_middleware(BrotliMiddleware)
+        # add middlewares
+        for middleware in self.middlewares:
+            self.app.add_middleware(middleware)