+++ stac_fastapi/extensions/stac_fastapi/extensions/core/__init__.py
@@ -3,6 +3,7 @@ from stac_fastapi.extensions.third_party.tiles import TilesExtension
 
 from .context import ContextExtension
 from .fields import FieldsExtension
+from .filter import FilterExtension
 from .query import QueryExtension
 from .sort import SortExtension
 from .transaction import TransactionExtension
@@ -10,6 +11,7 @@ from .transaction import TransactionExtension
 __all__ = (
     "ContextExtension",
     "FieldsExtension",
+    "FilterExtension",
     "QueryExtension",
     "SortExtension",
     "TilesExtension",