+++ stac_fastapi/extensions/stac_fastapi/extensions/core/fields.py
@@ -1,5 +1,5 @@
 """fields extension."""
-from typing import Set
+from typing import List, Optional, Set
 
 import attr
 from fastapi import FastAPI
@@ -20,21 +20,26 @@ class FieldsExtension(ApiExtension):
 
     Attributes:
         default_includes (set): defines the default set of included fields.
+        conformance_classes (list): Defines the list of conformance classes for the extension
 
     """
 
+    conformance_classes: List[str] = attr.ib(
+        factory=lambda: ["https://api.stacspec.org/v1.0.0-beta.3/item-search/#fields"]
+    )
+    schema_href: Optional[str] = attr.ib(default=None)
     default_includes: Set[str] = attr.ib(
-        default=attr.Factory(
-            lambda: {
-                "id",
-                "type",
-                "geometry",
-                "bbox",
-                "links",
-                "assets",
-                "properties.datetime",
-            }
-        )
+        factory=lambda: {
+            "id",
+            "type",
+            "stac_version",
+            "geometry",
+            "bbox",
+            "links",
+            "assets",
+            "properties.datetime",
+            "collection",
+        }
     )
 
     def register(self, app: FastAPI) -> None: