+++ stac_fastapi/pgstac/stac_fastapi/pgstac/core.py
@@ -1,18 +1,24 @@
 """Item crud client."""
 import re
 from datetime import datetime
-from typing import Any, Dict, List, Optional, Union
+from typing import Any, Dict, List, Optional, Type, Union
+from urllib.parse import urljoin
 
 import attr
 import orjson
+from asyncpg.exceptions import InvalidDatetimeFormatError
 from buildpg import render
+from fastapi import HTTPException
+from pydantic import ValidationError
+from stac_pydantic.links import Relations
+from stac_pydantic.shared import MimeTypes
 from starlette.requests import Request
 
 from stac_fastapi.pgstac.models.links import CollectionLinks, ItemLinks, PagingLinks
 from stac_fastapi.pgstac.types.search import PgstacSearch
 from stac_fastapi.types.core import AsyncBaseCoreClient
-from stac_fastapi.types.errors import NotFoundError
-from stac_fastapi.types.stac import Collection, Conformance, Item, ItemCollection
+from stac_fastapi.types.errors import InvalidQueryParameter, NotFoundError
+from stac_fastapi.types.stac import Collection, Collections, Item, ItemCollection
 
 NumType = Union[float, int]
 
@@ -21,19 +27,12 @@ NumType = Union[float, int]
 class CoreCrudClient(AsyncBaseCoreClient):
     """Client for core endpoints defined by stac."""
 
-    async def conformance(self, **kwargs) -> Conformance:
-        """Conformance classes."""
-        return Conformance(
-            conformsTo=[
-                "https://stacspec.org/STAC-api.html",
-                "http://docs.opengeospatial.org/is/17-069r3/17-069r3.html#ats_geojson",
-                "https://api.stacspec.org/v1.0.0-beta.2/item-search",
-            ]
-        )
+    search_request_model: Type[PgstacSearch] = attr.ib(init=False, default=PgstacSearch)
 
-    async def all_collections(self, **kwargs) -> List[Collection]:
+    async def all_collections(self, **kwargs) -> Collections:
         """Read all collections from the database."""
         request: Request = kwargs["request"]
+        base_url = str(request.base_url)
         pool = request.app.state.readpool
 
         async with pool.acquire() as conn:
@@ -50,7 +49,25 @@ class CoreCrudClient(AsyncBaseCoreClient):
                     collection_id=coll["id"], request=request
                 ).get_links()
                 linked_collections.append(coll)
-        return linked_collections
+        links = [
+            {
+                "rel": Relations.root.value,
+                "type": MimeTypes.json,
+                "href": base_url,
+            },
+            {
+                "rel": Relations.parent.value,
+                "type": MimeTypes.json,
+                "href": base_url,
+            },
+            {
+                "rel": Relations.self.value,
+                "type": MimeTypes.json,
+                "href": urljoin(base_url, "collections"),
+            },
+        ]
+        collection_list = Collections(collections=linked_collections or [], links=links)
+        return collection_list
 
     async def get_collection(self, id: str, **kwargs) -> Collection:
         """Get collection by id.
@@ -76,7 +93,7 @@ class CoreCrudClient(AsyncBaseCoreClient):
             )
             collection = await conn.fetchval(q, *p)
         if collection is None:
-            raise NotFoundError
+            raise NotFoundError(f"Collection {id} does not exist.")
         links = await CollectionLinks(collection_id=id, request=request).get_links()
         collection["links"] = links
         return Collection(**collection)
@@ -102,22 +119,26 @@ class CoreCrudClient(AsyncBaseCoreClient):
         # pool = kwargs["request"].app.state.readpool
         req = search_request.json(exclude_none=True)
 
-        async with pool.acquire() as conn:
-            q, p = render(
-                """
-                SELECT * FROM search(:req::text::jsonb);
-                """,
-                req=req,
+        try:
+            async with pool.acquire() as conn:
+                q, p = render(
+                    """
+                    SELECT * FROM search(:req::text::jsonb);
+                    """,
+                    req=req,
+                )
+                items = await conn.fetchval(q, *p)
+        except InvalidDatetimeFormatError:
+            raise InvalidQueryParameter(
+                f"Datetime parameter {search_request.datetime} is invalid."
             )
-            items = await conn.fetchval(q, *p)
+
         next: Optional[str] = items.pop("next", None)
         prev: Optional[str] = items.pop("prev", None)
         collection = ItemCollection(**items)
         cleaned_features: List[Item] = []
-        if collection["features"] is None or len(collection["features"]) == 0:
-            raise NotFoundError("No features found")
 
-        for feature in collection["features"]:
+        for feature in collection.get("features") or []:
             feature = Item(**feature)
             if (
                 search_request.fields.exclude is None
@@ -139,7 +160,8 @@ class CoreCrudClient(AsyncBaseCoreClient):
                 if include and len(include) == 0:
                     include = None
             cleaned_features.append(feature)
-            collection["features"] = cleaned_features
+
+        collection["features"] = cleaned_features
         collection["links"] = await PagingLinks(
             request=request,
             next=next,
@@ -148,7 +170,7 @@ class CoreCrudClient(AsyncBaseCoreClient):
         return collection
 
     async def item_collection(
-        self, id: str, limit: int = 10, token: str = None, **kwargs
+        self, id: str, limit: Optional[int] = None, token: str = None, **kwargs
     ) -> ItemCollection:
         """Get all items from a specific collection.
 
@@ -162,13 +184,16 @@ class CoreCrudClient(AsyncBaseCoreClient):
         Returns:
             An ItemCollection.
         """
-        req = PgstacSearch(collections=[id], limit=limit, token=token)
-        collection = await self._search_base(req, **kwargs)
+        # If collection does not exist, NotFoundError wil be raised
+        await self.get_collection(id, **kwargs)
+
+        req = self.search_request_model(collections=[id], limit=limit, token=token)
+        item_collection = await self._search_base(req, **kwargs)
         links = await CollectionLinks(
             collection_id=id, request=kwargs["request"]
-        ).get_links(extra_links=collection["links"])
-        collection["links"] = links
-        return collection
+        ).get_links(extra_links=item_collection["links"])
+        item_collection["links"] = links
+        return item_collection
 
     async def get_item(self, item_id: str, collection_id: str, **kwargs) -> Item:
         """Get item by id.
@@ -181,9 +206,19 @@ class CoreCrudClient(AsyncBaseCoreClient):
         Returns:
             Item.
         """
-        req = PgstacSearch(ids=[item_id], limit=1)
-        collection = await self._search_base(req, **kwargs)
-        return Item(**collection["features"][0])
+        # If collection does not exist, NotFoundError wil be raised
+        await self.get_collection(collection_id, **kwargs)
+
+        req = self.search_request_model(
+            ids=[item_id], collections=[collection_id], limit=1
+        )
+        item_collection = await self._search_base(req, **kwargs)
+        if not item_collection["features"]:
+            raise NotFoundError(
+                f"Item {item_id} in Collection {collection_id} does not exist."
+            )
+
+        return Item(**item_collection["features"][0])
 
     async def post_search(
         self, search_request: PgstacSearch, **kwargs
@@ -207,7 +242,7 @@ class CoreCrudClient(AsyncBaseCoreClient):
         ids: Optional[List[str]] = None,
         bbox: Optional[List[NumType]] = None,
         datetime: Optional[Union[str, datetime]] = None,
-        limit: Optional[int] = 10,
+        limit: Optional[int] = None,
         query: Optional[str] = None,
         token: Optional[str] = None,
         fields: Optional[List[str]] = None,
@@ -260,5 +295,8 @@ class CoreCrudClient(AsyncBaseCoreClient):
             base_args["fields"] = {"include": includes, "exclude": excludes}
 
         # Do the request
-        search_request = PgstacSearch(**base_args)
+        try:
+            search_request = self.search_request_model(**base_args)
+        except ValidationError:
+            raise HTTPException(status_code=400, detail="Invalid parameters provided")
         return await self.post_search(search_request, request=kwargs["request"])