+++ stac_fastapi/types/stac_fastapi/types/extension.py
@@ -1,5 +1,6 @@
 """base api extension."""
 import abc
+from typing import List, Optional
 
 import attr
 from fastapi import FastAPI
@@ -8,6 +9,9 @@ from fastapi import FastAPI
 @attr.s
 class ApiExtension(abc.ABC):
     """Abstract base class for defining API extensions."""
+
+    conformance_classes: List[str] = attr.ib(factory=list)
+    schema_href: Optional[str] = attr.ib(default=None)
 
     @abc.abstractmethod
     def register(self, app: FastAPI) -> None: